Bug Fixes and Consistency Issues Within NIMClient.cpp
Fixed an issue where the client would not allow the user to select another server (because the main loop would not return to the top).

Fixed an issue with the server list not refreshing if the current server selected was unavailable.

Added a section to properly display the list of servers and give the user a list of numbers to select the server they wish to challenge.

Various formatting/consistency changes.